+++ CHANGELOG.md
@@ -8,6 +8,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [0.1.1] - 2024-11-07
+
+### Added
+
+- DatFile: add `filter` method.
+- DatFile: add `reverse` method.
+
+### Fixed
+
+- Fix mashups after merge conflicts.
+
+
 ## [0.1.0] - 2024-11-05
 
 ### Added